requirements.txt
mysql-connector-python
fastapi
httptools
uvloop
uvicorn
Jinja2